actor_emul (chorer v0.1.0)

This module simulate an actor's localview during a globalview.

Functions

add_proc_localvars(P, V)

Add a variable to the local variables of the localview.

add_proc_mess_queue(P, M)

Add a message to the message queue of the process.

add_proc_spawnvars(P, V)

Add a variable to the spawn arguments of the localview.

del_proc_mess_queue(P, M)

Delete a message from the message queue of the process.

get_proc_data(P)

Get the all the data of the process.

get_proc_edge_info(P, E)

Get the informations of a localview's edge.

get_proc_edges(P)

Get all the localview's current edges.

get_proc_localvars(P)

Get the list of the local variables.

get_proc_mess_queue(P)

Get the message queue of the process.

get_proc_out_degree(P)

Get all the number of the localview's current edges.

proc_loop(Data)

Loop function to simulate a process.

set_proc_data(P, Data)

Set the all the data of the process.

use_proc_transition(P, E)

Use a transition of the localview.